% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{analyse_shadow_effects}
\alias{analyse_shadow_effects}
\title{Compare data before and after shadowing}
\usage{
analyse_shadow_effects(original, shadowed, variables = NULL)
}
\arguments{
\item{original}{Original data}

\item{shadowed}{Shadowed data}

\item{variables}{Variables to compare}
}
\value{
Summary of shadow effects
}
\description{
Compare data before and after shadowing
}
