% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_classical_error}
\alias{apply_classical_error}
\title{Apply classical measurement error}
\usage{
apply_classical_error(data, params)
}
\description{
Apply classical measurement error
}
\keyword{internal}
