% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_correlated_error}
\alias{apply_correlated_error}
\title{Apply correlated measurement error}
\usage{
apply_correlated_error(data, params)
}
\description{
Apply correlated measurement error
}
\keyword{internal}
