% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_differential_error}
\alias{apply_differential_error}
\title{Apply differential measurement error}
\usage{
apply_differential_error(data, params)
}
\description{
Apply differential measurement error
}
\keyword{internal}
