% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_misclassification}
\alias{apply_misclassification}
\title{Apply misclassification to binary variables}
\usage{
apply_misclassification(data, params)
}
\description{
Apply misclassification to binary variables
}
\keyword{internal}
