% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadow.censoring_shadow}
\alias{apply_shadow.censoring_shadow}
\title{Apply censoring shadow}
\usage{
\method{apply_shadow}{censoring_shadow}(data, shadow, ...)
}
\description{
Apply censoring shadow
}
