% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadow.combined_shadow}
\alias{apply_shadow.combined_shadow}
\title{Apply combined shadow}
\usage{
\method{apply_shadow}{combined_shadow}(data, shadow, ...)
}
\description{
Apply combined shadow
}
