% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadow.item_missingness_shadow}
\alias{apply_shadow.item_missingness_shadow}
\title{Apply item-level missingness shadow}
\usage{
\method{apply_shadow}{item_missingness_shadow}(data, shadow, ...)
}
\description{
Apply item-level missingness shadow
}
