% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadow.measurement_error_shadow}
\alias{apply_shadow.measurement_error_shadow}
\title{Apply measurement error shadow}
\usage{
\method{apply_shadow}{measurement_error_shadow}(data, shadow, ...)
}
\description{
Apply measurement error shadow
}
