% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadow.positivity_shadow}
\alias{apply_shadow.positivity_shadow}
\title{Apply positivity violation shadow}
\usage{
\method{apply_shadow}{positivity_shadow}(data, shadow, ...)
}
\description{
Apply positivity violation shadow
}
