% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadow}
\alias{apply_shadow}
\title{Apply a shadow to data}
\usage{
apply_shadow(data, shadow, ...)
}
\arguments{
\item{data}{Data frame to apply shadow to}

\item{shadow}{A shadow object created with create_shadow()}

\item{...}{Additional arguments passed to specific shadow methods}
}
\value{
Modified data frame
}
\description{
Apply a shadow to data
}
