% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadow.selection_shadow}
\alias{apply_shadow.selection_shadow}
\title{Apply selection bias shadow}
\usage{
\method{apply_shadow}{selection_shadow}(data, shadow, ...)
}
\description{
Apply selection bias shadow
}
