% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows-extended.R
\name{apply_shadow.truncation_shadow}
\alias{apply_shadow.truncation_shadow}
\title{Apply truncation shadow}
\usage{
\method{apply_shadow}{truncation_shadow}(data, shadow, ...)
}
\arguments{
\item{data}{Data frame to apply shadow to}

\item{shadow}{A truncation shadow object}

\item{...}{Additional arguments (currently unused)}
}
\value{
Data frame with truncation applied
}
\description{
Apply truncation shadow
}
