% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadows}
\alias{apply_shadows}
\title{Apply multiple shadows sequentially}
\usage{
apply_shadows(data, shadows, verbose = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{shadows}{List of shadow objects}

\item{verbose}{Logical, print progress?}
}
\value{
Data frame with all shadows applied
}
\description{
Apply multiple shadows sequentially
}
