% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{as.data.frame.margot_scenario}
\alias{as.data.frame.margot_scenario}
\title{Convert scenario to data frame for summary}
\usage{
\method{as.data.frame}{margot_scenario}(x, ...)
}
\arguments{
\item{x}{A margot_scenario object}

\item{...}{Additional arguments}
}
\value{
A data frame with scenario information
}
\description{
Convert scenario to data frame for summary
}
