% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{as.data.frame.margot_shadow}
\alias{as.data.frame.margot_shadow}
\title{Convert shadow to data frame for summary}
\usage{
\method{as.data.frame}{margot_shadow}(x, ...)
}
\arguments{
\item{x}{A margot_shadow object}

\item{...}{Additional arguments}
}
\value{
A data frame with shadow information
}
\description{
Convert shadow to data frame for summary
}
