% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{as_scenario}
\alias{as_scenario}
\title{Convert old-style scenario to S3}
\usage{
as_scenario(old_scenario)
}
\arguments{
\item{old_scenario}{List-based scenario from old interface}
}
\value{
A validated margot_scenario S3 object
}
\description{
Converts scenarios created with the old interface to proper S3 objects
}
