% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{as_shadow}
\alias{as_shadow}
\title{Convert objects to shadows}
\usage{
as_shadow(x, ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Additional arguments passed to methods}
}
\value{
A margot_shadow object
}
\description{
Convert objects to shadows
}
