% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{c.margot_shadow}
\alias{c.margot_shadow}
\title{Combine multiple shadows into a list}
\usage{
\method{c}{margot_shadow}(...)
}
\arguments{
\item{...}{Shadow objects to combine}
}
\value{
A list of validated shadow objects with class "shadow_list"
}
\description{
Combine multiple shadows into a list
}
