% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-monte-carlo.R
\name{calculate_performance_metrics}
\alias{calculate_performance_metrics}
\title{Calculate performance metrics from MC results}
\usage{
calculate_performance_metrics(results_df)
}
\description{
Calculate performance metrics from MC results
}
\keyword{internal}
