% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-distributions.R
\name{check_distribution}
\alias{check_distribution}
\title{Check if a variable follows a specified distribution}
\usage{
check_distribution(x, dist, plot = TRUE)
}
\arguments{
\item{x}{Vector of values}

\item{dist}{Distribution specification or family name}

\item{plot}{Logical, create diagnostic plot?}
}
\value{
List with test results
}
\description{
Check if a variable follows a specified distribution
}
