% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-core.R
\name{check_positivity}
\alias{check_positivity}
\title{Check positivity of treatment probabilities}
\usage{
check_positivity(treatment_probs, trt_name, time, min_prob = 0.01)
}
\description{
Check positivity of treatment probabilities
}
\keyword{internal}
