% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shift-interventions.R
\name{create_lmtp_wrapper}
\alias{create_lmtp_wrapper}
\title{Create LMTP-compatible wrapper for shift functions}
\usage{
create_lmtp_wrapper(shift_fn, time = 1)
}
\arguments{
\item{shift_fn}{A margot.sim shift function with (data, time, trt) signature}

\item{time}{The time/wave to use when calling the shift function}
}
\value{
Function with (data, trt) signature for LMTP compatibility
}
\description{
Wraps margot.sim shift functions to be compatible with LMTP-style
(data, trt) signature while preserving time information.
}
\examples{
# Create margot.sim shift function
shift_up <- create_shift_intervention(shift_amount = 1, min_value = 1, max_value = 7)

# Wrap for LMTP use
lmtp_shift <- create_lmtp_wrapper(shift_up, time = 1)
}
