% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-memory-utils.R
\name{create_memory_monitor}
\alias{create_memory_monitor}
\title{Monitor memory usage during simulation}
\usage{
create_memory_monitor(interval = 10, threshold = 1000)
}
\arguments{
\item{interval}{Check interval in seconds}

\item{threshold}{Memory threshold in MB to trigger warning}
}
\value{
Function that monitors memory
}
\description{
Monitor memory usage during simulation
}
