% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-diagnostics.R
\name{diagnose_all_shadows}
\alias{diagnose_all_shadows}
\title{Diagnose all shadowing effects}
\usage{
diagnose_all_shadows(data, shadowed_data, shadows)
}
\arguments{
\item{data}{Original data}

\item{shadowed_data}{Data after applying shadows}

\item{shadows}{List of applied shadows}
}
\value{
Comprehensive diagnostics
}
\description{
Diagnose all shadowing effects
}
