% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-core.R
\name{.default_sim_params}
\alias{.default_sim_params}
\title{Default simulation parameter values}
\usage{
.default_sim_params()
}
\description{
Default simulation parameter values
}
\keyword{internal}
