% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-core-exports.R
\name{.margot_core_classes}
\alias{.margot_core_classes}
\title{Exports for margot.core Package}
\usage{
.margot_core_classes()
}
\value{
Character vector of S3 class names that should move to margot.core
}
\description{
This file contains functions and classes that will eventually be moved
to the margot.core package. They are currently defined here but will
be imported from margot.core in the future.
}
\details{
The margotverse architecture splits functionality across multiple packages:
- margot.core: Core S3 classes and common utilities
- margot.sim: Simulation and shadow functionality
- margot.lmtp: LMTP estimation methods
- margot.grf: Generalized random forest methods
- margot.viz: Visualization functions
- margot.report: Reporting templates

This file helps track what will move to margot.core.
Get list of S3 classes for margot.core
}
\examples{
\dontrun{
# see what classes will move
.margot_core_classes()
}
}
\keyword{internal}
