% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-examples.R
\name{example_basic_simulation}
\alias{example_basic_simulation}
\title{Example: Basic simulation}
\usage{
example_basic_simulation()
}
\description{
Example: Basic simulation
}
\keyword{internal}
