% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-examples.R
\name{example_causal_censoring_bias}
\alias{example_causal_censoring_bias}
\title{Example: Causal effects under censoring}
\usage{
example_causal_censoring_bias()
}
\description{
Example: Causal effects under censoring
}
\keyword{internal}
