% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-examples.R
\name{example_censoring_mechanisms}
\alias{example_censoring_mechanisms}
\title{Example: Different censoring mechanisms}
\usage{
example_censoring_mechanisms()
}
\description{
Example: Different censoring mechanisms
}
\keyword{internal}
