% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-monte-carlo.R
\name{example_mc_measurement_error}
\alias{example_mc_measurement_error}
\title{Example: Evaluate estimator under measurement error}
\usage{
example_mc_measurement_error()
}
\description{
Example: Evaluate estimator under measurement error
}
