% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-examples.R
\name{example_multiple_outcomes}
\alias{example_multiple_outcomes}
\title{Example: Multiple correlated outcomes}
\usage{
example_multiple_outcomes()
}
\description{
Example: Multiple correlated outcomes
}
\keyword{internal}
