% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-examples.R
\name{example_sampling_weights}
\alias{example_sampling_weights}
\title{Example: Sampling weights}
\usage{
example_sampling_weights()
}
\description{
Example: Sampling weights
}
\keyword{internal}
