% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-distributions.R
\name{generate_from_dist.custom_distribution}
\alias{generate_from_dist.custom_distribution}
\title{Generate from custom distribution function}
\usage{
\method{generate_from_dist}{custom_distribution}(dist, n, linear_predictor = NULL)
}
\description{
Generate from custom distribution function
}
