% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{is_scenario}
\alias{is_scenario}
\title{Check if object is a scenario}
\usage{
is_scenario(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if object is a scenario
}
