% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{is_shadow}
\alias{is_shadow}
\title{Test if object is a shadow}
\usage{
is_shadow(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x is a margot_shadow
}
\description{
Test if object is a shadow
}
