% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{length.margot_scenario}
\alias{length.margot_scenario}
\title{Get the number of shadows in a scenario}
\usage{
\method{length}{margot_scenario}(x)
}
\arguments{
\item{x}{A margot_scenario object}
}
\value{
Integer number of shadows
}
\description{
Get the number of shadows in a scenario
}
