% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-core-exports.R
\name{margot_optional_deps}
\alias{margot_optional_deps}
\title{List heavy dependencies that should move to Suggests}
\usage{
margot_optional_deps()
}
\value{
Data frame of packages and their reasons for being optional
}
\description{
List heavy dependencies that should move to Suggests
}
\examples{
margot_optional_deps()
}
