% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-core-exports.R
\name{margot_options}
\alias{margot_options}
\title{Standard margotverse options}
\usage{
margot_options(..., .list = NULL)
}
\arguments{
\item{...}{Named options to set}

\item{.list}{List of named options to set}
}
\value{
If no arguments, returns list of current margot options
}
\description{
Get or set standard options used across margotverse packages
}
\examples{
# get all margot options
margot_options()

# set specific options
margot_options(verbose = TRUE, parallel = FALSE)

# reset to defaults
margot_options(.list = margot_option_defaults())
}
