% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-distributions.R
\name{margot_simulate_flex}
\alias{margot_simulate_flex}
\title{Enhanced margot_simulate with flexible distributions}
\usage{
margot_simulate_flex(
  n,
  waves,
  distributions = NULL,
  ...,
  validate_distributions = TRUE
)
}
\arguments{
\item{n}{Integer. Number of subjects to simulate}

\item{waves}{Integer. Number of measurement waves (time points)}

\item{distributions}{Distribution set or list of distributions}
}
\value{
Simulated data with specified distributions
}
\description{
Enhanced margot_simulate with flexible distributions
}
