% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{new_shadow_list}
\alias{new_shadow_list}
\title{Create a shadow list}
\usage{
new_shadow_list(...)
}
\arguments{
\item{...}{Shadow objects or lists of shadow objects}
}
\value{
An object of class "shadow_list"
}
\description{
Create a shadow list
}
