% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-monte-carlo.R
\name{plot.margot_mc_results}
\alias{plot.margot_mc_results}
\title{Plot Monte Carlo results}
\usage{
\method{plot}{margot_mc_results}(x, type = c("histogram", "qq", "trace", "bias"), ...)
}
\description{
Plot Monte Carlo results
}
