% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-diagnostics.R
\name{plot.margot_positivity_diagnostic}
\alias{plot.margot_positivity_diagnostic}
\title{Plot positivity diagnostics}
\usage{
\method{plot}{margot_positivity_diagnostic}(x, type = "propensity", ...)
}
\arguments{
\item{x}{margot_positivity_diagnostic object}

\item{type}{Type of plot: "propensity", "overlap", "balance"}

\item{...}{Additional arguments}
}
\description{
Plot positivity diagnostics
}
