% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scenarios.R
\name{plot.margot_scenario_comparison}
\alias{plot.margot_scenario_comparison}
\title{Plot scenario comparison}
\usage{
\method{plot}{margot_scenario_comparison}(x, estimand = "ate", ...)
}
\arguments{
\item{x}{A margot_scenario_comparison object}

\item{estimand}{Which estimand to plot (default "ate")}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot object
}
\description{
Plot scenario comparison
}
