% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{plot.margot_shadow}
\alias{plot.margot_shadow}
\title{Plot method for shadow objects}
\usage{
plot.margot_shadow(x, data = NULL, ...)
}
\arguments{
\item{x}{A margot_shadow object}

\item{data}{Optional data to show shadow effects on}

\item{...}{Additional arguments}
}
\value{
A plot (base R or ggplot2)
}
\description{
Plot method for shadow objects
}
