% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-causal.R
\name{print.margot_causal_sim}
\alias{print.margot_causal_sim}
\title{Print method for margot_causal_sim}
\usage{
\method{print}{margot_causal_sim}(x, ...)
}
\description{
Print method for margot_causal_sim
}
