% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-monte-carlo.R
\name{print.margot_mc_results}
\alias{print.margot_mc_results}
\title{Print method for Monte Carlo results}
\usage{
\method{print}{margot_mc_results}(x, digits = 4, ...)
}
\description{
Print method for Monte Carlo results
}
