% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-diagnostics.R
\name{print.margot_positivity_diagnostic}
\alias{print.margot_positivity_diagnostic}
\title{Print method for positivity diagnostics}
\usage{
\method{print}{margot_positivity_diagnostic}(x, ...)
}
\arguments{
\item{x}{margot_positivity_diagnostic object}

\item{...}{Additional arguments}
}
\description{
Print method for positivity diagnostics
}
