% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scenarios.R
\name{print.margot_scenario}
\alias{print.margot_scenario}
\title{Print method for scenario objects}
\usage{
\method{print}{margot_scenario}(x, ...)
}
\arguments{
\item{x}{A margot_scenario object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for scenario objects
}
