% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-monte-carlo.R
\name{print.mc_comparison}
\alias{print.mc_comparison}
\title{Print method for MC comparison}
\usage{
\method{print}{mc_comparison}(x, digits = 4, ...)
}
\description{
Print method for MC comparison
}
