% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{print.shadow_list}
\alias{print.shadow_list}
\title{Print method for shadow lists}
\usage{
\method{print}{shadow_list}(x, ...)
}
\arguments{
\item{x}{A shadow_list object}

\item{...}{Additional arguments (unused)}
}
\value{
The object invisibly
}
\description{
Print method for shadow lists
}
