% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadow-dependencies.R
\name{reorder_shadows}
\alias{reorder_shadows}
\title{Reorder shadows to respect dependencies}
\usage{
reorder_shadows(shadows)
}
\arguments{
\item{shadows}{List of shadow objects}
}
\value{
Reordered list of shadows
}
\description{
Reorder shadows to respect dependencies
}
