% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-examples.R
\name{run_all_examples}
\alias{run_all_examples}
\title{Run all margot examples}
\usage{
run_all_examples(pause = TRUE)
}
\arguments{
\item{pause}{Logical. Pause between examples?}
}
\description{
Runs all example functions to demonstrate package capabilities.
}
