% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-scenario-library-simple.R
\name{scenario_collection_simple}
\alias{scenario_collection_simple}
\title{Get simple scenario collection}
\usage{
scenario_collection_simple()
}
\value{
Named list of scenarios
}
\description{
Returns a simple collection of scenarios that work with standard
margot_simulate() output
}
