% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{[.margot_scenario}
\alias{[.margot_scenario}
\title{Extract shadows from a scenario}
\usage{
\method{[}{margot_scenario}(x, i, ...)
}
\arguments{
\item{x}{A margot_scenario object}

\item{i}{Index or indices of shadows to extract}

\item{...}{Additional arguments (unused)}
}
\value{
A shadow object or list of shadow objects
}
\description{
Extract shadows from a scenario
}
