% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-diagnostics.R
\name{summarize_positivity}
\alias{summarize_positivity}
\title{Summarize positivity diagnostics}
\usage{
summarize_positivity(data, treatment, covariates, treatment_vals)
}
\description{
Summarize positivity diagnostics
}
\keyword{internal}
