% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadow-dependencies.R
\name{update_shadow_params}
\alias{update_shadow_params}
\title{Update shadow parameters based on dependencies}
\usage{
update_shadow_params(shadow, applied_shadows = list())
}
\arguments{
\item{shadow}{Shadow object to update}

\item{applied_shadows}{List of already applied shadows}
}
\value{
Updated shadow object
}
\description{
Update shadow parameters based on dependencies
}
