% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-simulate-core.R
\name{validate_proportions}
\alias{validate_proportions}
\title{Validate that coefficients sum to less than 1}
\usage{
validate_proportions(coefs, node_name, threshold = 0.95)
}
\description{
Validate that coefficients sum to less than 1
}
\keyword{internal}
