% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_scenario.default}
\alias{validate_scenario.default}
\title{Default scenario validator}
\usage{
\method{validate_scenario}{default}(x, ...)
}
\description{
Default scenario validator
}
\keyword{internal}
