% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_scenario.margot_scenario}
\alias{validate_scenario.margot_scenario}
\title{Base scenario validator}
\usage{
\method{validate_scenario}{margot_scenario}(x, ...)
}
\description{
Base scenario validator
}
\keyword{internal}
