% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_scenario}
\alias{validate_scenario}
\title{Validator for margot_scenario objects}
\usage{
validate_scenario(x, ...)
}
\arguments{
\item{x}{An object to validate}

\item{...}{Additional arguments}
}
\value{
The validated object (invisibly)
}
\description{
Validator for margot_scenario objects
}
