% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.censoring_shadow}
\alias{validate_shadow.censoring_shadow}
\title{Censoring shadow validator}
\usage{
\method{validate_shadow}{censoring_shadow}(x, ...)
}
\description{
Censoring shadow validator
}
\keyword{internal}
