% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.coarsening_shadow}
\alias{validate_shadow.coarsening_shadow}
\title{Coarsening shadow validator}
\usage{
\method{validate_shadow}{coarsening_shadow}(x, ...)
}
\description{
Coarsening shadow validator
}
\keyword{internal}
