% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.default}
\alias{validate_shadow.default}
\title{Default shadow validator}
\usage{
\method{validate_shadow}{default}(x, ...)
}
\description{
Default shadow validator
}
\keyword{internal}
