% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.item_missingness_shadow}
\alias{validate_shadow.item_missingness_shadow}
\title{Item missingness shadow validator}
\usage{
\method{validate_shadow}{item_missingness_shadow}(x, ...)
}
\description{
Item missingness shadow validator
}
\keyword{internal}
