% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.margot_shadow}
\alias{validate_shadow.margot_shadow}
\title{Base shadow validator}
\usage{
\method{validate_shadow}{margot_shadow}(x, ...)
}
\description{
Base shadow validator
}
\keyword{internal}
