% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.measurement_error_shadow}
\alias{validate_shadow.measurement_error_shadow}
\title{Measurement error shadow validator}
\usage{
\method{validate_shadow}{measurement_error_shadow}(x, ...)
}
\description{
Measurement error shadow validator
}
\keyword{internal}
