% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.mode_effects_shadow}
\alias{validate_shadow.mode_effects_shadow}
\title{Mode effects shadow validator}
\usage{
\method{validate_shadow}{mode_effects_shadow}(x, ...)
}
\description{
Mode effects shadow validator
}
\keyword{internal}
