% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.positivity_shadow}
\alias{validate_shadow.positivity_shadow}
\title{Positivity shadow validator}
\usage{
\method{validate_shadow}{positivity_shadow}(x, ...)
}
\description{
Positivity shadow validator
}
\keyword{internal}
