% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.selection_shadow}
\alias{validate_shadow.selection_shadow}
\title{Selection shadow validator}
\usage{
\method{validate_shadow}{selection_shadow}(x, ...)
}
\description{
Selection shadow validator
}
\keyword{internal}
