% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{validate_shadow.truncation_shadow}
\alias{validate_shadow.truncation_shadow}
\title{Truncation shadow validator}
\usage{
\method{validate_shadow}{truncation_shadow}(x, ...)
}
\description{
Truncation shadow validator
}
\keyword{internal}
