% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadow-dependencies.R
\name{visualize_shadow_dependencies}
\alias{visualize_shadow_dependencies}
\title{Visualize shadow dependencies}
\usage{
visualize_shadow_dependencies(shadows = NULL)
}
\arguments{
\item{shadows}{List of shadow objects (optional)}
}
\value{
Prints dependency graph
}
\description{
Visualize shadow dependencies
}
