% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-suggests.R
\name{with_arrow}
\alias{with_arrow}
\title{Use arrow for large data operations}
\usage{
with_arrow(expr)
}
\arguments{
\item{expr}{Expression to evaluate with arrow}
}
\value{
Result of expression or NULL if arrow not available
}
\description{
Use arrow for large data operations
}
\keyword{internal}
