% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadows.R
\name{apply_shadows}
\alias{apply_shadows}
\title{Apply multiple shadows sequentially}
\usage{
apply_shadows(data, shadows, verbose = FALSE, preserve_truth = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{shadows}{List of shadow objects}

\item{verbose}{Logical, print progress?}

\item{preserve_truth}{Logical, whether to preserve original values (default FALSE)}
}
\value{
Data frame with all shadows applied
}
\description{
Apply multiple shadows sequentially
}
