% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{as_shadow}
\alias{as_shadow}
\title{Convert old-style shadow to S3}
\usage{
as_shadow(old_shadow)
}
\arguments{
\item{old_shadow}{List-based shadow from old interface}
}
\value{
A validated margot_shadow S3 object
}
\description{
Converts shadows created with the old interface to proper S3 objects
}
