% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-shadow-dependencies.R
\name{check_shadow_ordering}
\alias{check_shadow_ordering}
\title{Check shadow ordering for dependencies}
\usage{
check_shadow_ordering(shadows)
}
\arguments{
\item{shadows}{List of shadow objects}
}
\value{
Logical indicating if ordering respects dependencies
}
\description{
Check shadow ordering for dependencies
}
