% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{new_shadow}
\alias{new_shadow}
\title{Create a margot shadow object}
\usage{
new_shadow(type, params, ...)
}
\arguments{
\item{type}{Character string specifying the shadow type}

\item{params}{List of parameters specific to the shadow type}

\item{...}{Additional arguments passed to shadow-specific constructors}
}
\value{
An object of class "margot_shadow" and "{type}_shadow"
}
\description{
Create a margot shadow object
}
\examples{
# create a measurement error shadow
shadow <- new_shadow(
  type = "measurement_error",
  params = list(
    variables = "x",
    error_type = "classical",
    sigma = 0.5
  )
)
}
