% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margot-s3-methods.R
\name{summary.margot_shadow}
\alias{summary.margot_shadow}
\title{Summary method for margot objects}
\usage{
\method{summary}{margot_shadow}(object, ...)
}
\arguments{
\item{object}{A margot object}

\item{...}{Additional arguments}
}
\value{
Summary information
}
\description{
Summary method for margot objects
}
